Revert "Introduce support for multiple runtimes (#1243)"
This reverts commit 1719a61e6292464dd6e27d518e8a878e001e20f9.
