+++ Cargo.toml
@@ -20,9 +20,7 @@ members = [
     "gtest",
     "node",
     "pallets/*",
-    "pallets/aux/shift-session-manager",
-    "runtime/gear",
-    "runtime/vara",
+    "runtime",
     "utils/gear-runtime-test-cli",
     "utils/regression-analysis",
     "utils/junit-common",
@@ -31,8 +29,6 @@ members = [
     "utils/economic-checks",
     "utils/economic-checks/fuzz",
 ]
-
-exclude = ["pallets/aux"]
 
 [profile.release]
 panic = "unwind"
