+++ Makefile
@@ -71,14 +71,6 @@ node:
 node-release:
 	@ ./scripts/gear.sh build node --release
 
-.PHONY: vara
-vara:
-	@ ./scripts/gear.sh build node --no-default-features --features=vara-native,lazy-pages
-
-.PHONY: vara-release
-vara-release:
-	@ ./scripts/gear.sh build node --release --no-default-features --features=vara-native,lazy-pages
-
 # Check section
 .PHONY: check
 check: check-gear check-examples
@@ -233,10 +225,6 @@ gtest: init-js gear-test-release examples
 rtest: init-js node-release examples
 	@ ./scripts/gear.sh test rtest yamls="$(yamls)"
 
-.PHONY: rtest-vara
-rtest-vara: init-js vara-release examples
-	@ ./scripts/gear.sh test rtest yamls="$(yamls)"
-
 .PHONY: test-pallet
 test-pallet:
 	@ ./scripts/gear.sh test pallet
@@ -260,8 +248,5 @@ doc:
 
 .PHONY: fuzz
 fuzz:
-	@ ./scripts/gear.sh test fuzz $(target)
+	@ ./scripts/gear.sh test fuzz
 
-.PHONY: fuzz-vara
-fuzz-vara:
-	@ ./scripts/gear.sh test fuzz --features=vara-native,lazy-pages --no-default-features $(target)
