+++ lazy-pages/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2018"
 license = "GPL-3.0"
 homepage = "https://gear-tech.io"
 repository = "https://github.com/gear-tech/gear"
+rust-version = "1.56"
 
 [dependencies]
 log = "0.4.17"
