--- node/cli/Cargo.toml
@@ -1,78 +1,0 @@
-[package]
-name = "gear-cli"
-version = "0.1.0"
-authors = ["Gear Technologies"]
-edition = "2018"
-build = "build.rs"
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
-
-[dependencies]
-clap = { version = "3.2", features = ["derive"] }
-scale-info = { version = "2.1.1", default-features = false, features = ["derive"] }
-
-# Gear
-gear-node-primitives = { path = "../primitives" }
-gear-runtime-test-cli = { path = "../../utils/gear-runtime-test-cli" }
-service = { package = "gear-service", path = "../service", default-features = false, optional = true }
-
-# Gear Runtimes
-gear-runtime = { path = "../../runtime/gear", optional = true }
-vara-runtime = { path = "../../runtime/vara", optional = true }
-gear-runtime-common = { path = "../../runtime/common", default-features = false }
-
-# Substrate client
-sc-cli = { version = "0.10.0-dev", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable", features = [
-	"wasmtime",
-], optional = true }
-sc-service = { version = "0.10.0-dev", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable", features = [
-	"wasmtime",
-], optional = true }
-sc-client-api = { version = "4.0.0-dev", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
-
-# Substrate primitives
-sp-core = { version = "6.0.0", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
-sp-inherents = { version = "4.0.0-dev", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
-sp-keyring = { version = "6.0.0", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
-sp-runtime = { version = "6.0.0", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
-sp-timestamp = { version = "4.0.0-dev", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
-
-# Frame pallets
-pallet-transaction-payment = { version = "4.0.0-dev", default-features = false, git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
-
-# Substrate other (benchmarking etc)
-frame-system = { version = "4.0.0-dev", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
-frame-benchmarking = { version = "4.0.0-dev", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
-frame-benchmarking-cli = { version = "4.0.0-dev", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable", optional = true }
-try-runtime-cli = { version = "0.10.0-dev", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable", optional = true }
-
-[build-dependencies]
-substrate-build-script-utils = { version = "3.0.0", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
-
-[features]
-default = ["cli"]
-cli = [
-	"service",
-	"sc-cli",
-	"sc-service",
-	"frame-benchmarking-cli",
-	"try-runtime-cli",
-]
-gear-native = [
-	"service/gear-native",
-	"gear-runtime-test-cli/gear-native",
-	"gear-runtime",
-]
-vara-native = [
-	"service/vara-native",
-	"gear-runtime-test-cli/vara-native",
-	"vara-runtime",
-]
-lazy-pages = ["service/lazy-pages"]
-runtime-benchmarks = ["service/runtime-benchmarks"]
-debug-mode = ["service/debug-mode"]
-try-runtime = [
-	"service/try-runtime",
-	"try-runtime-cli",
-]
