+++ pallets/gear-scheduler/Cargo.toml
@@ -30,16 +30,16 @@ sp-std = { version = "4.0.0-dev", git = "https://github.com/gear-tech/substrate.
 sp-io = { version = "6.0.0", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable", default-features = false }
 
 [dev-dependencies]
-core-processor = { package = "gear-core-processor", path = "../../core-processor" }
-pallet-gear = { path = "../gear" }
-pallet-gear-messenger = { path = "../gear-messenger" }
-pallet-gear-program = { path = "../gear-program" }
-pallet-gear-gas = { path = "../gas" }
-sp-core = { version = "6.0.0", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
-sp-runtime = { version = "6.0.0", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
-pallet-balances = { version = "4.0.0-dev", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
-pallet-authorship = { version = "4.0.0-dev", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
-pallet-timestamp = { version = "4.0.0-dev", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
+core-processor = { package = "gear-core-processor", path = "../../core-processor", default-features = false }
+pallet-gear = { path = "../gear", default-features = false }
+pallet-gear-messenger = { path = "../gear-messenger", default-features = false }
+pallet-gear-program = { path = "../gear-program", default-features = false }
+pallet-gear-gas = { path = "../gas", default-features = false }
+sp-core = { version = "6.0.0", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable", default-features = false }
+sp-runtime = { version = "6.0.0", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable", default-features = false }
+pallet-balances = { version = "4.0.0-dev", default-features = false, git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
+pallet-authorship = { version = "4.0.0-dev", default-features = false, git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
+pallet-timestamp = { version = "4.0.0-dev", default-features = false, git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
 env_logger = "0.9"
 
 [features]
@@ -55,6 +55,14 @@ std = [
 	"sp-std/std",
 	"primitive-types/std",
 	"sp-io/std",
+	"sp-runtime/std",
+	"pallet-balances/std",
+	"pallet-authorship/std",
+	"pallet-timestamp/std",
+	"pallet-gear-gas/std",
+	"pallet-gear/std",
+	"pallet-gear-messenger/std",
+	"pallet-gear-program/std",
 ]
 runtime-benchmarks = [
 	"frame-benchmarking",
