+++ pallets/payment/Cargo.toml
@@ -59,6 +59,8 @@ std = [
 	"pallet-transaction-payment/std",
 	"pallet-authorship/std",
 	"primitive-types/std",
+	"pallet-gear-messenger/std",
+	"pallet-gear-scheduler/std",
 ]
 runtime-benchmarks = [
 	"frame-benchmarking",
