+++ runtime/Cargo.toml
@@ -1,9 +1,8 @@
 [package]
-name = "vara-runtime"
+name = "gear-runtime"
 version = "0.1.0"
 authors = ["Gear Technologies"]
 edition = "2018"
-build = "build.rs"
 license = "GPL-3.0"
 homepage = "https://gear-tech.io"
 repository = "https://github.com/gear-tech/gear"
@@ -15,27 +14,23 @@ targets = ["x86_64-unknown-linux-gnu"]
 codec = { package = "parity-scale-codec", version = "3.1.2", default-features = false, features = ["derive"] }
 log = { version = "0.4.17", default-features = false }
 scale-info = { version = "2.1.1", default-features = false, features = ["derive"] }
-static_assertions = "1.1.0"
 
-# Frame deps
+# Substrate deps
 frame-support = { version = "4.0.0-dev", default-features = false, git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
 frame-system = { version = "4.0.0-dev", default-features = false, git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
 frame-try-runtime = { version = "0.10.0-dev", default-features = false, git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable", optional = true }
 frame-executive = { version = "4.0.0-dev", default-features = false, git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
+pallet-aura = { version = "4.0.0-dev", default-features = false, git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
 pallet-authorship = { version = "4.0.0-dev", default-features = false, git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
-pallet-babe = { version = "4.0.0-dev", default-features = false, git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
 pallet-balances = { version = "4.0.0-dev", default-features = false, git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
 pallet-grandpa = { version = "4.0.0-dev", default-features = false, git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
-pallet-session = { version = "4.0.0-dev", default-features = false, git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
 pallet-sudo = { version = "4.0.0-dev", default-features = false, git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
 pallet-timestamp = { version = "4.0.0-dev", default-features = false, git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
 pallet-transaction-payment = { version = "4.0.0-dev", default-features = false, git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
 pallet-utility = { version = "4.0.0-dev", default-features = false, git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
-
-# Primitives
 sp-api = { version = "4.0.0-dev", default-features = false, git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
 sp-block-builder = { git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable", default-features = false, version = "4.0.0-dev" }
-sp-consensus-babe = { version = "0.10.0-dev", default-features = false, git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
+sp-consensus-aura = { version = "0.10.0-dev", default-features = false, git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
 sp-core = { version = "6.0.0", default-features = false, git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
 sp-inherents = { git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable", default-features = false, version = "4.0.0-dev" }
 sp-offchain = { version = "4.0.0-dev", default-features = false, git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
@@ -55,18 +50,15 @@ frame-system-benchmarking = { version = "4.0.0-dev", default-features = false, g
 hex-literal = { version = "0.3.4", optional = true }
 
 # Internal deps
-runtime-common = { package = "gear-runtime-common", version = "0.1.0", default-features = false, path = "../common" }
-pallet-gear-scheduler = { version = "1.0.0", default-features = false, path = "../../pallets/gear-scheduler" }
-pallet-gear-messenger = { version = "1.0.0", default-features = false, path = "../../pallets/gear-messenger" }
-pallet-gear-program = { version = "2.0.0", default-features = false, path = "../../pallets/gear-program" }
-pallet-gear = { version = "2.0.0", default-features = false, path = "../../pallets/gear" }
-pallet-gear-debug = { version = "2.0.0", default-features = false, path = "../../pallets/gear-debug", optional = true }
-pallet-gear-gas = { version = "2.0.0", default-features = false, path = "../../pallets/gas" }
-pallet-gear-payment = { version = "0.1.0", default-features = false, path = "../../pallets/payment" }
-pallet-gear-rpc-runtime-api = { version = "2.0.0", default-features = false, path = "../../pallets/gear/rpc/runtime-api" }
-gear-common = { version = "0.1.0", default-features = false, path = "../../common" }
-gear-node-primitives = { version = "1.0.0", default-features = false, path = "../../node/primitives" }
-pallet-shift-session-manager = { version = "0.1.0", default-features = false, path = "../../pallets/aux/shift-session-manager" }
+pallet-gear-scheduler = { version = "1.0.0", default-features = false, path = "../pallets/gear-scheduler" }
+pallet-gear-messenger = { version = "1.0.0", default-features = false, path = "../pallets/gear-messenger" }
+pallet-gear-program = { version = "2.0.0", default-features = false, path = "../pallets/gear-program" }
+pallet-gear = { version = "2.0.0", default-features = false, path = "../pallets/gear" }
+pallet-gear-debug = { version = "2.0.0", default-features = false, path = "../pallets/gear-debug", optional = true }
+pallet-gear-gas = { version = "2.0.0", default-features = false, path = "../pallets/gas" }
+pallet-gear-payment = { version = "0.1.0", default-features = false, path = "../pallets/payment" }
+pallet-gear-rpc-runtime-api = { version = "2.0.0", default-features = false, path = "../pallets/gear/rpc/runtime-api" }
+gear-common = { version = "0.1.0", default-features = false, path = "../common" }
 
 [build-dependencies]
 substrate-wasm-builder = { version = "5.0.0-dev", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
@@ -80,30 +72,27 @@ std = [
 	"frame-try-runtime/std",
 	"frame-system/std",
 	"frame-system-rpc-runtime-api/std",
-	"gear-common/std",
-	"gear-node-primitives/std",
+	"pallet-aura/std",
 	"pallet-authorship/std",
-	"pallet-babe/std",
 	"pallet-balances/std",
-	"pallet-gear-gas/std",
+	"pallet-gear-program/std",
+	"pallet-gear-scheduler/std",
 	"pallet-gear/std",
 	"pallet-gear-debug/std",
-	"pallet-gear-messenger/std",
+	"pallet-gear-gas/std",
 	"pallet-gear-payment/std",
-	"pallet-gear-program/std",
 	"pallet-gear-rpc-runtime-api/std",
+	"gear-common/std",
 	"pallet-grandpa/std",
-	"pallet-session/std",
 	"pallet-sudo/std",
 	"pallet-timestamp/std",
 	"pallet-transaction-payment/std",
 	"pallet-transaction-payment-rpc-runtime-api/std",
 	"pallet-utility/std",
-	"runtime-common/std",
 	"scale-info/std",
 	"sp-api/std",
 	"sp-block-builder/std",
-	"sp-consensus-babe/std",
+	"sp-consensus-aura/std",
 	"sp-core/std",
 	"sp-inherents/std",
 	"sp-offchain/std",
@@ -112,7 +101,6 @@ std = [
 	"sp-std/std",
 	"sp-transaction-pool/std",
 	"sp-version/std",
-	"pallet-shift-session-manager/std",
 ]
 runtime-benchmarks = [
 	"frame-benchmarking",
@@ -130,11 +118,8 @@ try-runtime = [
 	"frame-executive/try-runtime",
 	"frame-try-runtime",
 	"frame-system/try-runtime",
-	"pallet-authorship/try-runtime",
-	"pallet-babe/try-runtime",
 	"pallet-balances/try-runtime",
 	"pallet-grandpa/try-runtime",
-	"pallet-session/try-runtime",
 	"pallet-sudo/try-runtime",
 	"pallet-timestamp/try-runtime",
 	"pallet-transaction-payment/try-runtime",
