--- runtime/gear/src/constants.rs
@@ -1,47 +1,0 @@
-// This file is part of Gear.
-
-// Copyright (C) 2021-2022 Gear Technologies Inc.
-// SPDX-License-Identifier: GPL-3.0-or-later WITH Classpath-exception-2.0
-
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with this program. If not, see <https://www.gnu.org/licenses/>.
-
-/// Currency related constants
-pub mod currency {
-    use gear_node_primitives::Balance;
-
-    /// The existential deposit.
-    pub const EXISTENTIAL_DEPOSIT: Balance = 500;
-}
-
-/// Time and block constants
-pub mod time {
-    use gear_node_primitives::{BlockNumber, Moment};
-
-    /// This determines the average expected block time that we are targeting.
-    /// Blocks will be produced at a minimum duration defined by `SLOT_DURATION`.
-    /// `SLOT_DURATION` is picked up by `pallet_timestamp` which is in turn picked
-    /// up by `pallet_aura` to implement `fn slot_duration()`.
-    ///
-    /// Change this to adjust the block time.
-    pub const MILLISECS_PER_BLOCK: Moment = 1000;
-
-    // NOTE: Currently it is not possible to change the slot duration after the chain has started.
-    //       Attempting to do so will brick block production.
-    pub const SLOT_DURATION: Moment = MILLISECS_PER_BLOCK;
-
-    // Time is measured by number of blocks.
-    pub const MINUTES: BlockNumber = 60_000 / (MILLISECS_PER_BLOCK as BlockNumber);
-    pub const HOURS: BlockNumber = MINUTES * 60;
-    pub const DAYS: BlockNumber = HOURS * 24;
-}
