+++ scripts/check-spec.sh
@@ -11,7 +11,7 @@ check_spec() {
         if ($3 == "true")
             if ($1 > $2)
                 if ($1 - $2 >= 50 && $2 == 100)
-                    print "" # Ok case. We have reset the network
+                    print "" # Ok case. We have reseted the network
                 else
                     print "be bumped from",$1 # Should be bumped and be greater than master one
             else if ($1 == $2)
@@ -36,11 +36,10 @@ check_spec() {
 PACKAGES_REQUIRE_BUMP_SPEC="common core core-backend core-processor node pallets runtime-interface lazy-pages"
 
 SPEC_ON_MASTER="$(git diff origin/master | sed -n -r "s/^\-[[:space:]]+spec_version: +([0-9]+),$/\1/p")"
-ACTUAL_SPEC_GEAR="$(cat $ROOT_DIR/runtime/gear/src/lib.rs | grep "spec_version: " | awk -F " " '{print substr($2, 1, length($2)-1)}')"
-ACTUAL_SPEC_VARA="$(cat $ROOT_DIR/runtime/vara/src/lib.rs | grep "spec_version: " | awk -F " " '{print substr($2, 1, length($2)-1)}')"
+ACTUAL_SPEC="$(cat $ROOT_DIR/runtime/src/lib.rs | grep "spec_version: " | awk -F " " '{print substr($2, 1, length($2)-1)}')"
 
 if [ -z "$SPEC_ON_MASTER" ]; then
-    SPEC_ON_MASTER=$ACTUAL_SPEC_GEAR
+    SPEC_ON_MASTER=$ACTUAL_SPEC
 fi
 
 for package in $(git diff --name-only origin/master | grep ".rs$" | cut -d "/" -f1 | uniq); do
@@ -49,8 +48,4 @@ for package in $(git diff --name-only origin/master | grep ".rs$" | cut -d "/" -
     fi
 done
 
-echo "Checking spec for Gear runtime:"
-check_spec "$SPEC_ON_MASTER" "$ACTUAL_SPEC_GEAR" "$CHANGES"
-
-echo "Checking spec for Vara runtime:"
-check_spec "$SPEC_ON_MASTER" "$ACTUAL_SPEC_VARA" "$CHANGES"
+check_spec "$SPEC_ON_MASTER" "$ACTUAL_SPEC" "$CHANGES"
